Update version in pyproject.toml and add new source folder in monkeyFunctions.iml (#106)
Description: This commit updates the version in the pyproject.toml file from 0.1.1 to 0.1.3. Additionally, a new source folder has been added in the monkeyFunctions.iml file, pointing to the src directory. This change ensures that the project is using the correct version and includes all necessary source files for compilation and testing.

Co-authored-by: Jack Hopkins <jack@raveler.co.uk>