Feature/openai 1 3 5 (#109)
* Update dependencies in requirements.txt

Description:

* Commit subject: Fix OpenAI_API client initialization

Commit description:

* Commit subject line: Update openai version to 1.3.5

Commit description: This commit updates the openai dependency in the setup.cfg file from version 0.28.1 to version 1.3.5. The other dependencies remain unchanged. This update ensures compatibility with the latest version of openai and ensures a smooth functioning of the application. No other changes were made in this commit.

* small bugfixes, changed stringio to byteio

---------

Co-authored-by: Jack Hopkins <jack@raveler.co.uk>
Co-authored-by: Mart <mart@paperplane.ai>